Add duplicate check

+++ recipe.rb
@@ -12,18 +12,20 @@ git "dotfiles" do
   user "#{node[:user]}"
   destination "/home/#{node[:user]}"
   repository "https://github.com/taroooyan/dotfiles.git"
+  not_if "test -d /home/#{node[:user]}/dotfiles"
 end
-execute "dotfiles setup" do
-  user "#{node[:user]}"
-  cwd "/home/#{node[:user]}/dotfiles"
-  command "./setup.sh"
-end
+# execute "dotfiles setup" do
+#   user "#{node[:user]}"
+#   cwd "/home/#{node[:user]}/dotfiles"
+#   command "./setup.sh"
+# end
 
 package "vim"
 # NeoBundle
 directory "/home/#{node[:user]}/.vim/bundle" do
   user "#{node[:user]}"
   mode "755"
+  not_if "test -d /home/#{node[:user]}/.vim/bundle"
 end
 git "/home/#{node[:user]}" do
   user "#{node[:user]}"
@@ -40,10 +42,11 @@ execute "Install Go" do
   user "root"
   cwd "/tmp"
   command "chown -R root:root ./go; mv ./go /usr/local"
+  not_if "test -e /usr/local/go/bin/go"
 end
 
 execute "Install ghq" do
   user "#{node[:user]}"
-  command "go get github.com/motemen/ghq"
+  command "export GOPATH=$HOME; /usr/local/go/bin/go get github.com/motemen/ghq"
 end
